+++ src/main/java/org/PrintUtils.java
@@ -1,0 +1,20 @@
+package org;
+
+public class PrintUtils {
+
+    static String printBlue(String text) {
+        return ConsoleColors.BLUE + text + ConsoleColors.RESET;
+    }
+
+    static String printPurple(String text) {
+        return ConsoleColors.PURPLE + text + ConsoleColors.RESET;
+    }
+
+    static String printYellow(String text) {
+        return ConsoleColors.YELLOW + text + ConsoleColors.RESET;
+    }
+
+    static String printRed(String text) {
+        return ConsoleColors.RED + text + ConsoleColors.RESET;
+    }
+}
